backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pytz
python-dotenv
pika
sqlalchemy
alembic
psycopg2-binary
colorlog